profile component location 변경
