--- src/pages/profile/components/profile-data.tsx
@@ -1,12 +1,0 @@
-const ProfileData = () => {
-	return (
-		<div className='flex flex-col'>
-			<div> 닉네임, 친구추가 버튼, 메시지 버튼, 점 3개</div>
-			<div> 0경기 0승리 0패배 </div>
-			<div> intraId</div>
-			<div> bio</div>
-		</div>
-	);
-};
-
-export default ProfileData;
